Refactor(python) Move PEtab import code from command line script… (#825)
* Refactor(python) Move PEtab import code from command line script to AMICI library to make more reusable

* Install PEtab for testing

* Test PEtab import on travis

* Typehints

* make get_fixed_parameters work with dataframe only

* remove unnecessary lines from testMisc due to moving of petab_import
